First version of Get-HDCSDiags
Due to limitations and security restrictions by the infrastructure team I am not able to directly use invoke-command on remote computers from our company's windows 2012 servers. We are however able to run PSExec to launch the CMD file that is in the remote computer which starts the creation of diagnostic files. Once the process completes and the zip file is created it is then copied to the server we originally started the powershell tool from.
I have attempted to add some error handling but there are still some things that need to be added to polish off this tool. I will do so in the future and update this file when I am able to.