Create Task1.ps1
A department that deals with 2000+ computers needed to sign in to each computer and remove two files. I created a script that can delete the two files from each of the computers attached to dhcp server for each group of computers. It will save that department hours of doing a repetitive task.